# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestión de e-books, autores y categorías.\nUsa POST /api/v1/auth/login para obtener un token JWT y luego\ninclúyelo como `Authorization: Bearer \u003ctoken\u003e` en las peticiones protegidas.",
        "title": "E-Book Management API",
        "contact": {
            "name": "Soporte",
            "email": "soporte@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ebooks/{id}/cover": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sube una imagen de portada para un e-book existente y actualiza su registro.\nFormatos aceptados: JPG, JPEG, PNG, WebP. Tamaño máximo: 5 MB.\nRequiere autenticación JWT.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "Subir portada de e-book",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID del e-book",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Imagen de portada (JPG/PNG/WebP, máx 5 MB)",
                        "name": "cover",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UploadCoverResponse"
                        }
                    },
                    "400": {
                        "description": "UUID inválido, formato no soportado o archivo muy grande",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "E-book no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno o de storage",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ebooks/{id}/file": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sube un archivo de e-book (PDF, EPUB o MOBI) y lo asocia al e-book indicado.\nTamaño máximo: 50 MB. Requiere autenticación JWT.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "uploads"
                ],
                "summary": "Subir archivo de e-book",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID del e-book",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Archivo del libro (PDF/EPUB/MOBI, máx 50 MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "pdf",
                            "epub",
                            "mobi"
                        ],
                        "type": "string",
                        "description": "Formato del archivo",
                        "name": "format",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.UploadFileResponse"
                        }
                    },
                    "400": {
                        "description": "UUID inválido, formato no soportado o archivo muy grande",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "E-book no encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno o de storage",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.UploadCoverResponse": {
            "type": "object",
            "properties": {
                "cover_path": {
                    "description": "Path relativo del archivo en el bucket de Storage.\nÚsalo para construir la URL pública o una signed URL.\nexample: covers/550e8400-e29b-41d4-a716-446655440000.jpg",
                    "type": "string",
                    "example": "covers/550e8400-e29b-41d4-a716-446655440000.jpg"
                },
                "cover_url": {
                    "description": "URL pública de la portada (solo si el bucket es público).\nexample: https://xxxx.supabase.co/storage/v1/object/public/ebooks/covers/uuid.jpg",
                    "type": "string",
                    "example": "https://xxxx.supabase.co/storage/v1/object/public/ebooks/covers/uuid.jpg"
                }
            }
        },
        "handlers.UploadFileResponse": {
            "type": "object",
            "properties": {
                "file_path": {
                    "description": "Path relativo del archivo en el bucket de Storage.\nexample: pdf/550e8400-e29b-41d4-a716-446655440000.pdf",
                    "type": "string",
                    "example": "pdf/550e8400-e29b-41d4-a716-446655440000.pdf"
                },
                "file_url": {
                    "description": "URL pública del archivo (solo si el bucket es público).\nexample: https://xxxx.supabase.co/storage/v1/object/public/ebooks/pdf/uuid.pdf",
                    "type": "string",
                    "example": "https://xxxx.supabase.co/storage/v1/object/public/ebooks/pdf/uuid.pdf"
                },
                "format": {
                    "description": "Formato del archivo subido (epub, pdf o mobi).\nexample: pdf",
                    "type": "string",
                    "example": "pdf"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT obtenido de POST /api/v1/auth/login. Formato: `Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}